core: Add a KeycloakService implementation
The KeycloakService accesses the Keycloak REST API [1] and provides the
CRUD operations for groups, roles and users, which will be used later
for the ORT server endpoints. The KeycloakService can be configured
in the application.conf file, which contains by default the values from
the local docker compose environment.

In addition, there are tests for the KeycloakService covering the basic
operations, for which there is also a test-realm.json for the Keycloak test
container with example groups, roles and users.

[1]: https://www.keycloak.org/docs-api/19.0.1/rest-api/index.html

Signed-off-by: Onur Demirci <onur.demirci@bosch.io>