docs: Update customers database schema
As metadata of users, organizations and products will be stored in the
database, remove the Keycloak hint in the schema. Storing these values
in the database has the advantage that migrations of the models can be
simplified.
Model the relations accordingly.
While at it, replace the upper case column names with lower snake-case
notation as it is also used in the PostgreSQL documentation [1].

[1]: https://www.postgresql.org/docs/current/ddl-basics.html

Signed-off-by: Marcel Bochtler <marcel.bochtler@bosch.io>