Organizations: Introduce separate API objects for 'create' and 'update'
When creating an `Organization`, some non-nullable properties are not required
to be set by the API user, e.g. the `id` of the `Organization`.

Additionally, change the update endpoint from `PUT` to `PATCH`, to enable the
API user to update or delete specific properties of an entity in contrast to
only update the whole entity with a `PUT` request.
This change requires a helper type `OptionalValue` to be able to distinguish
between explicit `null` properties and `absent` properties.

Based on the API design decisions documented in
04777a0a9363e90eeca9db57da17b7255eb27a54.

Signed-off-by: Marcel Bochtler <marcel.bochtler@bosch.io>