dao: Migrate dao to model
Migrate the dao classes to implement the repositories from the model
module instead of the shared module. This also requires to adapt the
routes accordingly. This change completely decouples the dao module from
any API classes.

For the routes some type mappings have to be implemented which will
partly become obsolete and partly be moved to a separate module in a
later commit.

While at it, also align the tests a bit.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>