Move API classes to api-v1 module
Move all API classes and the related mappings to a new `api-v1` module.
Collecting API related code in a separate module ensures that if another
API version is introduced later, the different API versions are cleanly
separated and their code cannot mix up.

It might also be a good idea to move the actual route definitions to
`api-v1` as well, but this would require moving the authentication logic
to a separate module as well and is therefore left out for now.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>