Add a services layer
Move the business logic from the routes to service classes to separate
routing from business logic. The functions in the service classes are
grouped by the requirements of the services, not by entity like the
repositories.

While the services use the repository interfaces to be independent of
the actual implementation, they still depend on the database by creating
the transactions. It should be considered to decouple them completely
from the database by providing an abstraction for the transactions. It
also should be considered to remove the transactions from the DAO
implementations of the repositories and let the callers be responsible
for creating the transactions.

This commit also moves the mapping of database exceptions to a central
location. The reason is that exceptions are only thrown at the end of a
transaction, so we cannot rely anymore on them being thrown in the
repository classes.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>