model: Make the repository functions non-suspending
Remove the `suspend` modifier from the repository functions as a
preparation for creating the transactions outside the repositories. As
we currently use blocking transactions they cannot call suspending
functions.

This commit also changes the `dbQuery` function to be blocking to be
able to continue using it in the repositories. As the function is not
suspending anymore it also does not use `Dispatchers.IO` anymore, this
will be re-introduced in a later commit.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>