Add a repository for ORT runs
Add a repository for `OrtRun`s and use it in the `OrchestratorService`.
As part of this, switch from the currently unsued Java Time module for
Exposed to the Kotlinx Datetime module to be able to use a single
Datetime API throughout the source code.

As the job configuration class will change a lot during early
development store it as JSONB to not have to update the schema on each
change.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>