workers: Add configuration to build the Analyzer Docker image
This commit adds a new module `workers` and the configuration to build
the Docker image of the Analyzer worker step.
The image constructs a NPM package manager from ORT and calls
`npm --version`. This showcases how the future worker will call tools of
the Docker image programmatically through ORT's jars.

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.io>