docker-compose: Use host as network_mode for analyzer-worker
In order to reach the ort-server and keycloak services using the same address
as on the host machine, the analyzer-worker service has to use the same
network namespace by setting the network_mode to host. In this way, the ORT
server API (including authentication) can be used from the analyzer-worker
and the host machine. Otherwise, the ORT server API can be used either from
the worker container or from the host machine due to the meta-data set in the
access token (especially the `iss` property [1]).

[1] https://www.rfc-editor.org/rfc/rfc7519#section-4.1.1

Signed-off-by: Onur Demirci <onur.demirci@bosch.io>