Authentication: Use full property names of JWT configuration
Some `ApplicationConfiguration` implementations do not support getting child
nodes of the configuration.
For instance, `MapApplicationConfiguration`, which will be used in an upcoming
commit, does not offer a way to a add configurations hierarchically, and will
therefore fail to read the child node of the JWT configuration.
Switch to the full path of the configuration property to avoid these issues.

Signed-off-by: Marcel Bochtler <marcel.bochtler@bosch.io>