Gradle: Add a "transport:spi" module
The purpose of this module is to define basic interfaces serving as an
abstraction layer over concrete communication mechanisms how
messages are transported between services. With such an abstraction
layer in place, it should be possible to deploy the server components
in different environments.

"SPI" stands for "Service Provider Interface" and refers to the fact
that many of the interfaces declared here must be implemented by other
modules providing specific functionality. In the future, further
modules will be added that implement specific transport mechanisms.
Possible examples are direct HTTP calls, different messaging systems,
etc.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>