Model: Add classes representing messages to the Orchestrator
The classes define the API of the Orchestrator component. They
represent the messages to trigger a new analysis run and the results
that can be produced by the single workers.

Since the workers are not yet implemented, the message classes are
incomplete, and also their concrete payload still has to be defined.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>