Analyzer: Remove unused dependencies
The SW360 client and the Gradle tooling API are not required as
dependencies for the analyzer module. These are transitive dependencies
by the ORT analyzer.
Because these dependencies are not available on maven central, the
repositories have to be defined anyways for Gradle to be able to
download the metadata.

Signed-off-by: Marcel Bochtler <marcel.bochtler@bosch.io>