model: Add all needed model classes for an analyzer run
The model classes for an analyzer run are needed for the mapping from the dao
classes to business model classes. There is not a business model class for
every dao class, because some represent a map or list with primitive data
types. The reason why the ORT model classes were not used is:
1. To have a stable API and not to adapt the database with every change in ORT
2. To access the database entries somehow, their IDs have to be stored.

Signed-off-by: Onur Demirci <onur.demirci@bosch.io>