Kubernetes: Add configuration of Keycloak service
When sending a request to the Ort Server service, one has to
authenticate through Keycloak. Unfortunately, the address exposed by
Minikube does not match the one in the cluster (keycloak:8081). This
causes a JWT validation error because the issuer of the token is not the
one configured in the server.
The solution is to have a service of type 'LoadBalancer', expose it to
localhost with `minikube tunnel --bind-address=127.0.0.1` and to create
an alias for it to 'keycloak' in the hosts file.

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.io>