dao: Simplify creation of foreign keys
Use the short syntax to define foreign keys and rely on the default key
names generated by the database to shorten the migration scripts. As
part of this also always be explicit about the nullability of columns
and align the column names for foreign keys to use `_id` as suffix.

Also fix some foreign key column types which wrongly used serials.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>