dao: Rename options table to package_manager_configuration_options
Even though options will also be required in other places like the
scanner configuration, make the options table specific to the package
manager configurations. This is easier to maintain because the table
will only ever have one foreign key and any changes will be local to the
package manager configurations.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>