AnalyzeRequest: Use only the ID of the representing objects
Since the `AnalyzerWorker` has direct access to the database, it is no longer
necessary to have the whole object in the request, it is sufficient if it only
contains the IDs. Due to this change, the `AnalyzerWorker` has to be adjusted
in order to query the objects from the database.

Signed-off-by: Onur Demirci <onur.demirci@bosch.io>