Kubernetes: Set up permission for the orchestrator Pod
By default, the orchestrator Pod does not have access to the Kubernetes API to
do various tasks. Therefore, the orchestrator was not able to use the Kubernetes
transport abstraction implementation to create jobs. To allow the orchestrator
Pod to access the Kubernetes API a `ServiceAccount`, `Role` and `RoleBinding`
have to be added to the manifest.

Signed-off-by: Onur Demirci <onur.demirci@bosch.io>