dao: Remove package curations
Remove all tables and the related models that are related to package
curations. The ORT CLI applies package curations in the analyzer command
because they are required by later commands like the scanner and there
is no dedicated command to apply curations, and the ORT result files are
supposed to contain all data shared between the CLI commands. This does
not fit the server workflow very well where curations can be applied on
demand. Therefore remove the curations for now to re-introduce them
later in a way that better works for the server approach.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>