model: Remove id from classes which are not aggregate roots
Remove the `id` property from all model classes which do not represent
aggregate roots. They do not need an identifier because they cannot be
accessed independently from their aggregate root, and they should not
have an identifier because they are required when creating an instance
of their aggregate root, but their id cannot be known at this time.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>