dao: Fix nullable package manager configuration options
The `PackageManagerConfiguration.options` property is a nullable map.
Currently, the `PackageManagerConfigurationsTable` references options
from the `PackageManagerConfigurationOptionsTable` but this does not
allow to make the distinction between an empty map and a null one. The
problem is that there is a semantic difference between the two: If the
options are not null, they override global options. However if they are
null, global options are kept.
This commit fix this by adding a separate boolean column "has_options"
to represent the tri state.

Fixes OSMIDEV-674.

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.io>