KubernetesMessageSenderFactoryTest: Decouple from local config
The `KubernetesMessageSenderFactory` uses the
`ClientBuild.defaultClient()` function to create the Kubernetes client.
If no `KUBECONFIG` environment variable is set, this function reads the
`~/.kube/config` file if it exists which couples the test to the local
configuration. If `~/.kube/config` does not contain a valid
configuration, for example because it belongs to a currently not running
Minikube installation, creating the client fails with an exception.

Decouple the test from the local Kubernetes configuration by providing a
configuration file and setting the `KUBECONFIG` environment variable.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>