refactor(analyzer): Use helper functionality from transport:spi module
Make use of the EndpointComponent base class to simplify message
handling, configuration, and testing of the interactions with the
transport. This aligns the module with the Orchestrator
implementation.

In addition, remove the now obsolete AnalyzerReceiver class. It has
been replaced by the message handler function in the new
AnalyzerComponent class.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>