Helm: Change the Keycloak hostname for ORT Server Core
Since the move to Minikube, ORT Server Core couldn't contact Keycloak
because Bitnami chart prefixes the hostname with the namespace.
This commit fixes the Helm template for ORT Server Core. The alias to
localhost must unfortunately also be adapted or the iss field in the JWT
will cause the token to be rejected by the ORT Server.

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.io>