Compose: Add a second queue for communication with the Analyzer Worker
The Orchestrator creates a Message Receiver that can only process an
`OrchestratorMessage`. If the same queue is used to send requests to
the Analyzer Worker, the Orchestrator service will also receive these
messages and will fail with a Deserialization exception because the
messages are of type `AnalyzerRequest`.
This commit introduces a second queue for the Orchestrator/Analyzer
worker communication in the development Docker Compose setup.

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.io>