feat(kubernetes): Support mapping of environment variables
When setting up the environment for the job to create,
KubernetesMessageSender removes prefixes in environment variables that
match the target endpoint. This makes it possible to have variables
for specific workers without the risk of name clashes.

This mechanism could also be extended later to propagate only specific
environment variables to the worker.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>