feat(dao): Add functions to apply list query parameters
Add extension functions for ListQueryParameters and EntityClass that
incorporate the parameters into database queries. This makes it easy
to support these standard parameters in the database layer, for
instance in repository implementations.

Note that invalid column names for the order definition are not yet
handled gracefully; this is going to be fixed in a later commit.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>