fix(rabbitmq): Do not create consumers in an endless loop
There was a misunderstanding about the way the basicConsume() function
works: It does not receive a single message, but registers a consumer
that is called asynchronously when new messages arrive. Thus, the
function must not be called in a loop, since on each iteration another
consumer is registered. This behavior eventually killed the RabbitMQ
service.

Instead, register the consumer once and then block the current thread
until it is canceled (which should not happen in normal operation
mode).

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>