refactor(dao): Remove reference options from table definitions
Remove `ReferenceOption.CASCADE` from all table definitions. In most
cases it was wrong because the table definitions in SQL did not set `ON
DELETE CASCADE` for the foreign keys. In cases where it was correct it
was also meaningless to set the option because Exposed uses it only when
creating tables in the database which is a feature we do not use.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>