refactor(secrets): Remove listPaths() function from SecretsProvider
This function is problematic, since it assumes a hierarchical
organization of secrets - which is not guaranteed to be supported by
all possible secret storage implementations. In addition, this
functionality is also not needed by the ORT Server use cases, since
queries on secrets will be done on their metadata which is stored in
the database.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>