refactor(orchestrator): Move job scheduling to the end of functions
Move the job scheduling to the end of the functions to prepare for
running the job creation in a transaction before scheduling the jobs.

To simplify this, introduce a class to collect the created jobs, as the
handler functions can create different jobs depending on which jobs are
configured to run.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>