feat(orchestrator): Use transactions when handling messages
Use transactions when reading, updating, and creating jobs. This ensures
that the orchestrator handles jobs consistently. This also improves
error handling, as several possible exceptions where not caught before.

This is just a first step, the error handling needs to be further
improved to ensure that the job statuses are updated correctly. Also,
this does not yet introduce error recovery for scheduling new jobs.

While at it, also align the log output of the handler functions.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>