feat(orchestrator): Set the transaction isolation level to serializable
When handling jobs, it is necessary to avoid phantom reads and
serialization anomalies [1] to make sure that the correct jobs are
created and scheduled. Therefore set the transaction isolation level to
"serializable" [2], to ensure that no parallel transactions modify the
jobs jobs at the same time.

Note that while the documentation claims that exposed uses the isolation
level "repeatable read" by default [3], it actually uses "read
committed" for PostgreSQL [4].

[1]: https://www.postgresql.org/docs/current/transaction-iso.html
[2]: https://wiki.postgresql.org/wiki/Serializable
[3]: https://github.com/JetBrains/Exposed/wiki/Transactions#advanced-parameters-and-usage
[4]: https://github.com/JetBrains/Exposed/blob/0.41.1/exposed-core/src/main/kotlin/org/jetbrains/exposed/sql/Database.kt#L209-L214

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>