feat(kubernetes): Exclude specific variables when creating a job
The environment of the new job is initialized from the environment of
the current pod. Some variables, however, may cause problems when
copied directly. For instance, the PATH variable could override the
path definition from the Docker build. Therefore, drop a number of
(currently hard-coded) variables which may be problematic.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>