feat(vault): Add more configuration properties
Support a path prefix and a namespace. Path prefixes are required to
handle a secret engine that is configured at a different path than the
standard one.

Some environments use the Vault enterprise feature of namespaces to
support multiple independent tenants. Allowing the configuration of a
namespace is a prerequisite to support such environments.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>