feat(vault): Improve token handling
The existing implementation obtained a token once at startup and then
used it forever. Since tokens have a limited lifetime, this will not
work. Therefore, add a mechanism to obtain a new token if the current
one is rejected.

Unfortunately, the standard KTor bearer authentication plugin cannot
be used for this purpose, because Vault has some specifics not
supported by it.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>