feat(kubernetes): Add the traceId as label when creating a job
This is useful for diagnostic purposes, and also when mapping back a
failed job to the original ORT run.

Since the trace ID is an arbitrary string (typically a UUID), it can
be longer than the maximum size allowed for a label value. Therefore,
it is split if necessary and stored in multiple labels with a common
prefix.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>