fix(evaluator): Use the resolved revision for the OrtResult
When constructing the OrtResult to be passed to the evaluator, so far
the revision from the OrtRun entity was used. This causes failures
when determining relative paths for projects, which is based on VCS
information using the resolved revision.

To fix this for now, obtain the resolved revision from the top-level
project. This is only a work-around, however; a correct implementation
has to be created later that also handles submodules correctly.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>