fix(kubernetes): Detect a stalled watch iterator in JobWatchHelper
This is an attempt to fix a corner case that occurred in the watcher
component of the Kubernetes job monitor: Sometimes the monitor gets a
watch iterator that does not yield relevant events. In this case,
watching is started again and again with the same resource version.

To fix this, record the resource version that was used to create the
last watch iterator. If the next iterator is to be created with the
same resource version, watching has obviously stalled. Then obtain a
fresh resource version by querying the current list of jobs.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>