feat(keycloak): Improve type-safety by introducing value classes
The Keycloak API sometimes requires entity names and sometimes entity
ids. This can make it easy to confuse those values. For example, the API
to add a sub group to a group requires the id of the parent group but
the name of the sub group. Introduce value classes for entity ids and
names to make the API of the client unambiguous and reduce the risk for
mistakes.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>