test(keycloak): Create the test realm programmatically
Create the test realm programmatically using the Keycloak admin client
instead of importing a JSON file. This makes it easier to keep the realm
in sync with the test data and reduces the amount of configuration
required to create the test realm to the minimum required for the tests.

Note that with the new approach it would be possible to recreate the
realm for each test case, but this would increase the duration of the
`KeycloakClientTest` by about a minute.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>