test(dao): Work around a transaction manager leak
If the Exposed default database is changed by calling
`Database.connect()`, it can happen that the first transaction on the
new connection still uses the transaction manager from the previous
connection. This could happen if multiple tests using the
`DatabaseTestExtension` were executed.

Work around this issue by making sure that any still open transaction
manager is closed before connecting to the new database.

Another solution would be to not rely on the default database and
instead manage the database instance, but this would requier a bigger
refactoring of the `dao` module.

[1]: https://github.com/JetBrains/Exposed/issues/454

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>