fix(dao): Do not use the Exposed default database
Always require a dedicated `Database` instance instead of relying on
Exposed's default database, because using the default database can lead
to issues when replacing it with a new one by calling
`Database.connect()`.

These issues were partly addressed in 5bd498b by resetting the
transaction manager, but this solved the issue only for the thread in
which the database was initialized. It could still happen that the
`ThreadLocal` used by `TransactionManager` to manage transaction
manager instances held references to transaction managers that belong
to the wrong database.

In particular this was an issue when using `mockkTransaction` and
`DatabaseTestExtension` in the same test module and both used
`Dispatchers.IO` for running their transactions. In this case it could
happen that one of the IO threads was associated with a
`NotInitializedManager` even though the `DatabaseTestExtension` did set
the correct default database, resulting in exceptions being thrown.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>