refactor(dao): Use the `each` instead of the `test` functions
Use `beforeEach` and `afterEach` instead of `beforeTest` and
`afterTest`. The difference is that the `each` functions run only for
tests with the type `TestType.Test` while the `test` functions also run
for tests with the type `TestType.Container` [1]. However, there is
currently no situation where those functions need to be executed for
the containers so using the `test` functions introduces unnecessary
overhead. For consistency, also apply this change for test classes that
do not use containers.

[1]: https://kotest.io/docs/framework/lifecycle-hooks.html

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>