refactor(dao): Remove the `fixture` parameter of `DatabaseTestExtension`
Remove the workaround to provide a `fixture` parameter to the
`DatabaseTestExtension` which is then executed in the `beforeEach`
lifecycle callback and instead provide a `fixtures` object as a property
of the `DatabaseTestExtension` class. This provides more convenience for
tests as they can use the `Fixtures` from the extension and do not have
to create their own.

The issue that motivated the workaround has been further investigated
and it turned out that it only occurs if the `suspend fun beforeEach`
function is overridden by the test class, therefore document that the
`beforeEach {}` DSL function must be used instead.

A Kotest issue [1] has been created to clarify if the situation can be
improved.

While at it, also align the formatting of test classes a bit.

[1]: https://github.com/kotest/kotest/issues/3555

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>