feat(workers): Introduce EnvironmentService
This service is going to provide functionality related to setting up
the build environment based on user configuration. The initial
implementation has a function to find an infrastructure service that
matches a repository URL. Based on this, the credentials for the
repository to check out before the analyzer runs are determined.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>