ci: Run integration tests in a separate job
Run the integration tests in a separate job to improve the overall CI
performance. As currently the `core` module contains all integration
tests and no unit tests, just run the tests for this module in a
separate job. This could later be improved by using either test tags [1]
or separate source sets.

[1]: https://kotest.io/docs/framework/tags.html

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>