feat(services): Add a function to synchronize roles and groups
Add a function to `AuthorizationService` to synchronize the roles based
on the role enums and database entities with the Keycloak roles and
groups. The function also ensures that all groups have the correct roles
assigned and that all roles have the correct composite roles assigned.

While at it, also add documentation for some previously introduced
private helper functions in `DefaultAuthorizationService`.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>