fix(services): Fix a potential issue with role synchronization
Move the final underscore separator to the `rolePrefix` functions to
prevent a potential issue when synchronizing roles. The prefix is used
for looking up unneeded roles, without the final underscore valid roles
could be identified as unneeded. For example, for the prefix
`permission_organization_1` the role `permission_organization_10` could
have wrongly been identified as unneeded.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>