refactor(keycloak)!: Do not return HTTP responses
Change the return type for functions that do not return a value from
`HttpResponse` to `Unit`. The `HttpResponse` is an implementation detail
that should not be exposed to the users. Also, those functions throw an
exception in case of an error, so there is no need to return anything in
case of success. This also allows to remove the Ktor client from the
public API of the Gradle module.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>