deps: Upgrade kotest-extension-testcontainers to 2.0.2
For the release notes see [1].

Switch to `JdbcDatabaseContainerExtension` because
`JdbcTestContainerExtension` is deprecated. This requires to slightly
modify the way the data source is created for the tests.

Also, use the new `ContainerExtension` instead of the deprecated
`TestContainerExtension`. The new extension class calls the `configure`
closure after starting the container instead of before. This causes
issues with the port mapping in the `ArtemisTestContainer`, so make sure
that it is configured before it is installed.

[1]: https://github.com/kotest/kotest-extensions-testcontainers#200

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>