refactor(workers): Move EnvironmentConfiguration to workers:common
This class is only used to set up the environment for specific
workers. Therefore, it does not need to be located in the central
model package.

While at it, shorten its name (the term "Config" is preferred in this
project) and add a property for the environment definitions required
by the current repository.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>