feat: Add a superuser role and group
Add a superuser role that grants all permissions and a superuser group
to make it easy to manage the superusers. This will be taken into use in
an upcoming commit that introduces authorization to the API endpoints.

Note that the superuser role does not have any composite roles assigned
to simplify the implementation. Instead it will always override the
requirements for any other permission roles.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>