feat: Add superuser role and group to the Keycloak realm
Add the superuser role and group to the `master-realm.json` which is
used in the Docker Compose setup and add the `ort-admin` user to the
newly created group. This simplifies testing the API with the
`ort-admin` user once authorization is implemented because the user
already has access to all resources.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>