refactor(core): Rewrite `AuthenticationIntegrationTest` to use Keycloak
Rewrite the `AuthenticationIntegrationTest` to use Keycloak instead of a
mock server for generating and verifying the JWT tokens. This is
required, because an upcoming change will introduce fetching data from
Keycloak as part of the authentication.

The new implementation also has the benefit that it verifies not only
that the handling of the JWT token is correct, but also that the actual
integration with Keycloak works correctly.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>