refactor(keycloak): Do not set `expectSuccess` to `true` by default
Do not set the HTTP client property `expectSuccess` to `true` by default
when configuring authentication. The setting is expected when using the
`DefaultKeycloakClient`, but is not helpful when using the helper
function in integration tests. The reason is that failed requests are
causing exceptions with this setting which breaks the usual test pattern
to verify the exit code after executing a request.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>