feat(Analyzer): Support an environment configuration from the job
Check in AnalyzerWorker whether the current job configuration contains
an environment configuration. If so, set up the build environment
based on this configuration; otherwise, try reading the environment
configuration file from the repository.

This makes it possible to use an environment configuration even if the
repository to be analyzed does not define one; also, an existing
environment configuration in the repository could be overwritten.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>