refactor: Split `:services` into separate modules
Split the `:services` module into separate Gradle modules for better
separation of concerns and to reduce the amount of transitive
dependencies a consumer inherits that uses only part of the services.

While at it, make sure that all dependencies that are used in the public
API of the services are marked as API dependencies.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>