feat(reporter): Add an `OrtResultReporter`
Add a reporter that creates an ORT result file as a report. This is
useful for comparing the results of an ORT run with ORT results created
in other environments than the server.

Ideally this would have been implemented as an API endpoint that
generates the ORT result file on the fly, but this is currently not
possible, because the ORT server API relies on kotlinx.serialization
while the ORT result file can only be created with Jackson.

A possible solution would be to create the ORT result on the disk first,
but this is not very efficient, potentially uses up the disk space of
the core module for large results, and also would require to add a lot
of transitive dependencies to the core module increasing its size.

Therefore it was decided to go with the reporter approach for now. This
might be reconsidered once [1] is implemented.

[1]: https://github.com/oss-review-toolkit/ort/issues/3904

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>