feat(core): Add and configure the `MicrometerMetrics` plugin
This commit adds the required dependencies for the `MicrometerMetrics`
plugin and the used `GraphiteMeterRegistry` [1]. Additionally, it
configures the plugin and adds it to the `Application`. The
`GraphiteMeterRegistry` can be configured via the application.conf file.
By default, it uses the `PLAINTEXT` protocol and its corresponding port
to push data to `Graphite`.

[1]: https://micrometer.io/docs/registry/graphite
[2]: https://graphite-web.readthedocs.io/en/latest/feeding-carbon.html#the-plaintext-protocol

Signed-off-by: Onur Demirci <onur.demirci@bosch.io>