feat(core): Add a prefix to the metrics of `MicrometerMetrics`
By adding a prefix to all metrics, it's possible to distinguish from which
system the metrics are coming [1]. This is helpful when the Graphite
instance is used from several systems. By default the prefix is
configured as `ort-server-core`, but it can be changed by setting the
environment variable `GRAPHITE_PREFIX`.

[1]: https://micrometer.io/docs/registry/graphite#_prefixing_your_metrics

Signed-off-by: Onur Demirci <onur.demirci@bosch.io>