feat(api): Include job details in ORT run responses
Include details about all created jobs in the ORT runs returned by
endpoints in the repositories route. This provides the user with
detailed information about the current status of an ORT run.

Note that only the `OrtRun` class in the `api-v1` module is extended to
include the job details, the `OrtRun` class of the `model` module does
not need that information as the jobs should only be loaded on demand.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>