feat(dao): Add properties to OrtRun related to configuration
The new `configContext` property can be specified when creating a new
run to select a specific set of configuration properties. It is
going to be evaluated by the new Config worker, which will also
resolve the value and store it in the `resolvedConfigContext`
property.

Add also a new `issues` property that is populated by the Config
worker if it detects problems with the configuration.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>