refactor(config): Change the API of ConfigManager
Do not pass a `Context` anymore when creating an instance. Instead,
pass the context when calling functions related to configuration
files.

This change is necessary because a resolved context is not always
available when creating the ConfigManager. It is only relevant when
dealing with configuration files; so other usages of ConfigManager
should not have to deal with it.

Unfortunately, this change caused many adaptions on client classes.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>