feat(config): Allow reading secrets from the application config
In deployments without a dedicated secret storage, it is often useful
to treat secrets analogously to normal configuration properties (and
pass them for instance as environment variables). To support such
setups, ConfigManager now offers a mode in which secrets are first
looked up via the configuration; only if they cannot be found there,
the secret provider is queried. This mode is enabled by default, but
can be turned off.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>