refactor(kubernetes): Add a separate config for job monitor
So far, the configuration properties for the Kubernetes job monitor
component were placed in the application.conf of the transport-spi
module. This does not really make sense. The component will need its
specific configuration in the future anyway to access secrets.

So, move the properties related to the job monitor to a dedicated
application.conf file for this module. Also, add a runtime dependency
to the currently available ConfigSecretProvider implementation.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>