refactor: Rename the job config related properties
Rename the properties related to job configurations to include "job" in
their names. This is a preparation for storing the resolved
configuration for an ORT run [1] to avoid naming conflicts and to better
separate job configurations which are input to the jobs from the config
that is resolved within the jobs (for example package curations).

[1]: https://github.com/oss-review-toolkit/ort/blob/15e33b2310cb9eacd464121a325dd12ae8b173f1/model/src/main/kotlin/OrtResult.kt#L71-L75

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>