feat: Support syntax highlighting for parameter validation script
Change the script compilation configuration for the validation scripts
used by ConfigWorker to support syntax highlighting and
auto-completion in the IDE. To achieve this, the script extension is
changed to "params.kts", since the association is done via the
extension. Also, a marker file has to be added to the META-INF folder.

These changes use ORT as example where highlighting of scripts is
working. Background information can be found at [1].

[1]: https://discuss.kotlinlang.org/t/make-auto-completion-work-in-kotlin-scripts-with-context/10185/19

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>